backend/requirements.txt
asgiref==3.7.2
Django~=3.2.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
